minify files and change Validation method into public
