+++ server/pkg/accounts/accounts.go
@@ -53,7 +53,7 @@ func CreateAccount(db database.Database, logger log.Logger, account *models.Acco
 		return fmt.Errorf("an error occured salting the account password: %s", err.Error())
 	}
 
-	err = account.validateAccount()
+	err = account.ValidateAccount()
 	if err != nil {
 		return fmt.Errorf("account could not be validated: %s", err.Error())
 	}
@@ -69,7 +69,8 @@ func CreateAccount(db database.Database, logger log.Logger, account *models.Acco
 	return nil
 }
 
-func (account models.Account) validateAccount() error {
+//Validate account method
+func (account models.Account) ValidateAccount() error {
 	return validation.ValidateStruct(&account,
 		validation.Field(&account.Username, validation.Required, is.Alphanumeric, validation.Min(3)),
 		validation.Field(&account.Email, validation.Required, is.Email),
