--- server/pkg/projects/validator.go
@@ -1,13 +1,0 @@
-package projects
-
-import (
-	"github.com/darwinfroese/hacksite/server/models"
-	validation "github.com/go-ozzo/ozzo-validation"
-)
-
-// ValidateProject checks if the model is valid
-func (project models.Project) ValidateProject() error {
-	return validation.ValidateStruct(&project,
-		validation.Field(&project.Name, validation.Required),
-	)
-}
